requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djoser==2.2.3
pillow==10.3.0
python-dotenv==1.0.1
drf-spectacular==0.27.2
django-phonenumber-field[phonenumbers]==7.3.0